+++ logic/pkg/models/user.go
@@ -8,7 +8,7 @@ import (
 
 // User represents the userinfo table in the database
 type User struct {
-	ID          uint   `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID          uint64 `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name        string `gorm:"column:name" json:"name"`
 	Spw         string `gorm:"column:spw" json:"spw"`
 	Creation    string `gorm:"column:creation" json:"creation"`
@@ -21,7 +21,7 @@ type User struct {
 
 // UserInfo represents the userinfo that is safe to return to the client
 type UserInfo struct {
-	ID         uint   `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID         uint64 `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name       string `gorm:"column:name" json:"name"`
 	Creation   string `gorm:"column:creation" json:"creation"`
 	Product    string `gorm:"column:product" json:"product"`
