+++ syncing/models/refreshKey.go
@@ -5,9 +5,9 @@ import (
 )
 
 type RefreshKey struct {
-	ID             uint      `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID             uint64    `gorm:"primaryKey;autoIncrement" json:"id"`
 	RefreshKey     string    `gorm:"column:refresh_key;not null" json:"refreshKey"`
-	UserID         uint      `gorm:"column:user_id;not null" json:"userId"`
+	UserID         uint64    `gorm:"column:user_id;not null" json:"userId"`
 	ExpirationTime time.Time `gorm:"column:expiration_time;type:timestamp" json:"expirationTime"`
 	UsageCount     int       `gorm:"column:usage_count;not null;default:0" json:"usageCount"`
 }
