+++ syncing/models/sessionKey.go
@@ -5,9 +5,9 @@ import (
 )
 
 type SessionKey struct {
-	ID             uint      `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID             uint64    `gorm:"primaryKey;autoIncrement" json:"id"`
 	SessionKey     string    `gorm:"column:session_key;not null" json:"sessionKey"`
-	UserID         uint      `gorm:"column:user_id;not null" json:"userid"`
+	UserID         uint64    `gorm:"column:user_id;not null" json:"userid"`
 	ExpirationTime time.Time `gorm:"column:expiration_time;type:timestamp" json:"expirationTime"`
 	CreationTime   time.Time `gorm:"column:creation_time;type:timestamp with time zone;default:current_timestamp" json:"creationTime"`
 	Valid          string    `gorm:"column:valid;type:varchar(1);default:'0'" json:"valid"`
